CODE-RUB: Code cleanup
